Communicate changes to a task the user is editing via flash messages
The user should have a stable view of a task while editing it. I.e., if the task
is changed from another client, we don't want to automatically update the user's
form with the new values or something like that -- that's disruptive and super
annoying.
By the same token, we don't want the user to inadvertantly overwrite someone
else's changes.

Since the is-complete checkbox is not part of the edit form, we notify the user
of the change in completedness status but auto-dismiss the notification after a
delay. If additional such changes come in, we immediately dismiss the
notification for the latest one -- there's no sense in subjecting the user to a
stream of notifications about the same thing where each one is displayed for its
full duration.

On the other hand, if a change modifies more than the completedness of a task
that the user is currently editing, then the user needs to decide whether to
* cancel editing, allowing them to view the changed task (and potentially then
make an edit), or
* continue editing and overwrite the recent change on saving.
In this case, I leave the notification displayed until either,
* the user selects and action from the notification, or
* the user cancels or completes editing with the form controls.