Fix taskDiff fn by adding optional arg of keys to diff on
Each input element is capable of changing some subset of a task's fields. Only
that input element knows what that subset is so it should be able to specify the
fields.

Ex:
There is a checkbox in the UI to mark a task as complete or not.
The "output" of that checkbox is
  {isComplete: true|false}
We only want to compare this output to the task it relates to on the key
"isComplete".

Why not just merge the output of a control into the original task and then diff
that against a task, using all fields?
Fields can be cleared. We may remove the description, for example.