Fix editing a task that has been deleted
I had not worked this out all the way. If the user was editing a task that was
deleted by another client, then receiving the deletion would cause the task to
be removed from tasks.tasksById in the Redux store, which would mean that the
task id held in tasks.taskToEdit would point to nothing. The edit form would
then be destroyed in the subsequent re-render and the user would NOT have the
choice of whether or not to continue editing the task.

Now editing the task is actually not disrupted by the task being deleted in the
database. Much like deleting a file on disk doesn't kill a text-editor buffer,
deleting the file in the database doesn't mean that the user loses the
information from the form. And just as a buffer can be re-written to a new file,
the task information can be re-written to a new task.